+++ scanpipe/pipes/__init__.py
@@ -272,6 +272,9 @@ def run_command(cmd, log_output=False):
 def remove_prefix(text, prefix):
     """
     Remove the `prefix` from `text`.
+    Note that build-in `removeprefix` was added in Python3.9 but we need to keep
+    this one for Python3.8 support.
+    https://docs.python.org/3.9/library/stdtypes.html#str.removeprefix
     """
     if text.startswith(prefix):
         prefix_len = len(prefix)