+++ scanpipe/tests/test_pipes.py
@@ -1149,14 +1149,14 @@ class ScanPipePipesTest(TestCase):
         p1 = Project.objects.create(name="Analysis")
         package = pipes.update_or_create_package(p1, package_data1)
         self.assertEqual("pkg:deb/debian/adduser@3.118?arch=all", package.purl)
-        self.assertEqual("", package.primary_language)
+        self.assertEqual("bash", package.primary_language)
         self.assertEqual(datetime.date(1999, 10, 10), package.release_date)
 
         updated_data = dict(package_data1)
-        updated_data["primary_language"] = "Python"
+        updated_data["notice_text"] = "NOTICE"
         updated_package = pipes.update_or_create_package(p1, updated_data)
         self.assertEqual("pkg:deb/debian/adduser@3.118?arch=all", updated_package.purl)
-        self.assertEqual("Python", updated_package.primary_language)
+        self.assertEqual("NOTICE", updated_package.notice_text)
         self.assertEqual(package.pk, updated_package.pk)
 
         resource1 = CodebaseResource.objects.create(project=p1, path="filename.ext")