+++ CHANGELOG.rst
@@ -11,6 +11,10 @@ v33.0.0 (unreleased)
   a Project.
   https://github.com/nexB/scancode.io/issues/583
 
+- Add fields in CycloneDX BOM output using the component properties.
+  See registered properties at https://github.com/nexB/aboutcode-cyclonedx-taxonomy
+  https://github.com/nexB/scancode.io/issues/637
+
 - Refine the "Command Line Interface" documentation about the ``scanpipe`` command
   usages in the Docker context.
   Add the /app workdir in the "PYTHONPATH" env of the Docker file to make the